+++ python_modules/dagster/dagster/_api/snapshot_execution_plan.py
@@ -7,8 +7,8 @@ from dagster._core.definitions.events import AssetKey
 from dagster._core.errors import DagsterUserCodeProcessError
 from dagster._core.execution.plan.state import KnownExecutionState
 from dagster._core.instance import DagsterInstance
+from dagster._core.remote_origin import RemoteJobOrigin
 from dagster._core.remote_representation.external_data import DEFAULT_MODE_NAME
-from dagster._core.remote_representation.origin import RemoteJobOrigin
 from dagster._core.snap.execution_plan_snapshot import (
     ExecutionPlanSnapshot,
     ExecutionPlanSnapshotErrorData,
