+++ python_modules/dagster/dagster/_core/remote_representation/grpc_server_registry.py
@@ -8,10 +8,7 @@ from typing_extensions import TypeGuard
 import dagster._check as check
 from dagster._core.errors import DagsterUserCodeProcessError, DagsterUserCodeUnreachableError
 from dagster._core.instance import InstanceRef
-from dagster._core.remote_representation.origin import (
-    CodeLocationOrigin,
-    ManagedGrpcPythonEnvCodeLocationOrigin,
-)
+from dagster._core.remote_origin import CodeLocationOrigin, ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.types.loadable_target_origin import LoadableTargetOrigin
 from dagster._time import get_current_timestamp
 from dagster._utils.error import SerializableErrorInfo, serializable_error_info_from_exc_info
