+++ python_modules/dagster/dagster/_grpc/types.py
@@ -13,14 +13,10 @@ from dagster._core.execution.plan.state import KnownExecutionState
 from dagster._core.execution.retries import RetryMode
 from dagster._core.instance.ref import InstanceRef
 from dagster._core.origin import JobPythonOrigin, get_python_environment_entry_point
+from dagster._core.remote_origin import CodeLocationOrigin, RemoteJobOrigin, RemoteRepositoryOrigin
 from dagster._core.remote_representation.external_data import (
     DEFAULT_MODE_NAME,
     job_name_for_partition_set_snap_name,
-)
-from dagster._core.remote_representation.origin import (
-    CodeLocationOrigin,
-    RemoteJobOrigin,
-    RemoteRepositoryOrigin,
 )
 from dagster._serdes import serialize_value, whitelist_for_serdes
 from dagster._utils.error import SerializableErrorInfo
