+++ python_modules/dagster/dagster_tests/daemon_sensor_tests/test_sensor_run.py
@@ -27,13 +27,13 @@ from dagster._core.definitions.sensor_definition import (
 )
 from dagster._core.events import DagsterEventType
 from dagster._core.log_manager import LOG_RECORD_METADATA_ATTR
+from dagster._core.remote_origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.remote_representation import (
     RemoteInstigatorOrigin,
     RemoteRepositoryOrigin,
     RemoteSensor,
 )
 from dagster._core.remote_representation.external import RemoteRepository
-from dagster._core.remote_representation.origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.scheduler.instigation import (
     DynamicPartitionsRequestResult,
     InstigatorState,
