+++ python_modules/dagster/dagster_tests/daemon_tests/conftest.py
@@ -6,12 +6,12 @@ from typing import Optional, cast
 
 import dagster as dg
 import pytest
+from dagster._core.remote_origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.remote_representation import (
     CodeLocation,
     InProcessCodeLocationOrigin,
     RemoteRepository,
 )
-from dagster._core.remote_representation.origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.test_utils import (
     InProcessTestWorkspaceLoadTarget,
     create_test_daemon_workspace_context,
