+++ python_modules/dagster/dagster_tests/scheduler_tests/test_scheduler_run.py
@@ -12,6 +12,7 @@ import pytest
 from dagster import AssetExecutionContext, AssetKey, DefaultScheduleStatus, schedule
 from dagster._core.definitions.run_request import RunRequest
 from dagster._core.instance import DagsterInstance
+from dagster._core.remote_origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.remote_representation import (
     CodeLocation,
     GrpcServerCodeLocation,
@@ -20,7 +21,6 @@ from dagster._core.remote_representation import (
     RemoteRepositoryOrigin,
 )
 from dagster._core.remote_representation.external import RemoteRepository, RemoteSchedule
-from dagster._core.remote_representation.origin import ManagedGrpcPythonEnvCodeLocationOrigin
 from dagster._core.scheduler.instigation import (
     InstigatorState,
     InstigatorStatus,
