+++ python_modules/libraries/dagster-aws/dagster_aws_tests/ecs_tests/launcher_tests/test_utils.py
@@ -1,4 +1,4 @@
-from dagster._core.remote_representation.origin import (
+from dagster._core.remote_origin import (
     RegisteredCodeLocationOrigin,
     RemoteJobOrigin,
     RemoteRepositoryOrigin,
