+++ python_modules/libraries/dagster-postgres/dagster_postgres_tests/compat_tests/test_back_compat.py
@@ -1057,7 +1057,7 @@ def test_add_backfill_tags(hostname, conn_string):
 
 
 def test_add_bulk_actions_job_name_column(hostname, conn_string):
-    from dagster._core.remote_representation.origin import (
+    from dagster._core.remote_origin import (
         GrpcServerCodeLocationOrigin,
         RemotePartitionSetOrigin,
         RemoteRepositoryOrigin,
